+++ cli/src/main/kotlin/com/specificlanguages/mops/cli/FindOperations.kt
@@ -7,7 +7,12 @@ import picocli.CommandLine.ParentCommand
 @Command(
     name = "find",
     description = ["Search editable MPS project sources."],
-    subcommands = [FindInstancesCommand::class, FindUsagesCommand::class, RecursiveHelpCommand::class],
+    subcommands = [
+        FindInstancesCommand::class,
+        FindUsagesCommand::class,
+        FindByNameCommand::class,
+        RecursiveHelpCommand::class,
+    ],
 )
 class FindOperations : CliCommand() {
     @ParentCommand